Various fixes related to or discovered when fixing SVY-7028: Refactor/implement custom JSON types
There were some problems with webservices code client side (it didn't
initialize smart properties correctly) + when setting a full value on
the client when you had no value defined for it before (and thus no
conversion info) would make it reach the server as-is.

To fix that we could either refactor that type conversion info is sent
once in the beginning for services/components and not sent any more
attached to data only, but for now, the server side can handle such a
case and send back to client-side a notification to "instrument" the new
value - so watch it and so on.