For SVY-10350 : Crediteuren screen sometimes is blank and solution is not responsive
First 2 fixes for this case.

1. A requestFocus() (delayed, async) was called on a comp. of a form
that is not shown yet, then a getWidth() gets called on another form. It
hung because getWidth() was waiting to get results on server side but on
client it waited for requestFocus() to execute first. Request focus on
client forced a touch of the needed form to be able to load it in hidden
div, but server would not reply to that because it was blocked by
getWidth. This fix bumps force touch event level so that it executes
anyway.

2. Api calls were always executed in a timeout after form state was
resolved. But that is wrong, because by the time the timeout executes
some forms are already hidden/unresolved => api not found... And the API
is present if checked without timeout.

There is one more problem that lead to this hang (which now is no longer
reproducivle due to the fix above) but I will commit that separately.