SVY-10413 deadlock when closing in active clients at the same time as starting a new one
removed all the sync blocks and use a concurrent hashmap for the
sessions instead of using synchronize blocks.

closeSessions are guarded by a atomic boolean so that only 1 thread can
check the sessions, if others are wanting to do  it at the same time
they just skip it, there is no need to do it right after each other all
the time. They can wait for the next one