SVY-10186 SPIKE: improve lazy loading of component's controller SVY-9947 Selected tab in tabpanel is not persisted if the tabs were not set at design time.
make sure that the obj.msg (the forms data) are pushed before any calls
are made so that the data is up to date.
Don't use getFormState() for that because that waits for the form to be
initialized and then it initalizes with old data.
Because of this removeWatches() is not always possible (form is not
really there) and the scope is not there yet for the converters.