For: SVY-10964, SVY-11137, SVY-10932, SVY-11027, SVY-10931
Fixed an unrelated bug specific to - 8.2 noticed while testing the
merged changes for these cases. Array type when sending an 'addition'
(for example new column in extra table) should write the whole value of
that column to json, not the 'update', because there is no previous
value for that addition on client.