For SVY-10992 : Error when removing columns from extra table
Alternate fix for the problem in last commit (the unit tests failure).
Existing code was supposed to handle that properly but because if used
List.contains (that uses .equals) to check if an element is in the list
it would give false positives - if the element was an empty map and we
had other empty maps in the array.