For SVY-11409 : Servoy extra table - selecting "-none-" for foundset in designer does not work
Fixed many problems, not just the original one in the case.
The idea is that now property default values are not always "null"...
# properties that are not converted to persists (so not custom object,
array or component prop.) should now work properly with default values
from .spec file; I tested this with extra-table foundset prop. with
default value and without default value in spec:
    - restore to default (so right click in properties view on the
*name* of the property then restore to default) should really
restore/show in the UI the default from .spec. In the .frm file that
prop. should disappear (not be stored - as it is the default)
    - when default is shown in UI, setting it to none/null or any other
value that is not default should work and save in the .frm file the
newly selected value (null or whatever it is)
    - when changing the value manually to the default value and saving
it should disappear from the UI.
    - same tests as the above adapted to when there is no default value
in .spec - it should work properly in this case as well
# same as all the above but with properties/subproperties that do
convert to persists mixed with ones that don't (arrays of custom
objects, custom objects directly as root properties, custom objects with
arrays (of ints or other custom objects), ...). All this should work
properly both when there is a default value in .spec for them and when
it is not. (see and adapt all subpoints above)
# custom objects directly as component property or in arrays of custom
objects could not be set to -none- (no matter the default value); also
handling none for those gave all kinds of exceptions; the eraser button
in properties view would just blank out the sub-properties instead of
removing it/set to null