For SVY-14172: combination servoyextra-listformcomponent and showFormPopup has issues
Problem here is that the foundset property type would not be able to
send changes to server when the component wants that (loadExtraRecords
for example) while executing in foundset property type listener which is
actually triggered by a server-to-client conversion that changes
property's reference (it was a full foundset value comming from server
to replace previous one).

So when loadExtraRecords was called, the new foundset property value's
changeNotifier was not yet set (so it could not notify that it has stuff
to send back to server). And also if foundset property type would check
itseld if it already has changes when setChangeNotifier is called on it,
that change notifier would still work with the old value's reference
which was wrong (it got the old value that was still in the model at
that point).