For SVY-14140: Custom types - containing Date fields are not converted correctly
Initial server side implementation for types being sent to client per
component/service as needed (+ dynamic types optimized in viewports) as
opposed to how it worked previously (writing all the time the type for
any data that needs client side conversion and not being aware of the
type client side if data was null or for api param. types and so on).

It should also improve a lot the web-socket traffic - at least for
forms/components that use a lot of client-side converted data - besides
offering support for typed retvals/params.