For SVY-14140: Custom types - containing Date fields are not converted correctly
Partial commit of client side impl. just in order to be able to merge
master into client_side_types again (as the master commits are stacking
up).

The basics and a few important types are already refactored to be aware
of client side types through the types registry, but more still need to
be refactored.