SVY-15329 Add "Client Status" information to clientmanager plugin
don't use getEndpoint() because that wants a synchronized lock.
For getting last ping time that is not needed, if it is not there then
it is 0
But getLastPingTime can be called from the ClientManager service or from
the admin page. So that can potentially deadlock if the client already
has the window lock