For SVYX-431: NavBar doesn't work after client reconnects
When wifi is closed for example, but restored after a few seconds, the
server should ask the client what forms it already has so that it only
sends updates when wsSession.onOpen is called instead of a full value.

Problem in this scenario was that server thought it needs to fully
resend a form (which increases an array/custom obj. version), although
that was present already on client; and client ignored that full update
(so client version for that array/obj remained the old one). Then they
would be out-of-sync in an unexpected way, although it didn't actually
change. Not on client, not on server.