For SVY-14140
Do it similar to what code did before (full send of array tabs from
client to server didn't work correctly because client used oldValue
undefined and server used oldValue something that could be unrelated
anyway after a full update). I created a couple of follow-ups to make it
more as it should be: 17855 and 17854.