For SVY-18721: element attributes property cannot be modified in place
When "attributes" which is ServoyAttributesPropertyType has a design
value and then in the onload of the form a new attribute was added, code
in RhinoConverter for the default object conversion that
ServoyAttributesPropertyType uses (by not having a conversion) did mark
the whole attributes as being changed to be sent to client - but that
attributes map was in the "defaultAndTemplate" section of the
BasicWebObject - and only stuff in "properties" are sent to client; so
it would send a 'null' which was wrong.

Plain 'object' types do not have this as they can't be set from
properties view afaik.